+++ src/Humans.Web/Controllers/ApplicationController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Localization;
 using Humans.Application.Interfaces;
+using Humans.Application.Interfaces.Governance;
 using Humans.Domain.Constants;
 using Humans.Domain.Enums;
 using Humans.Web.Authorization;
