+++ tests/Humans.Application.Tests/Services/ApplicationDecisionServiceTests.cs
@@ -5,7 +5,8 @@ using NodaTime;
 using NodaTime.Testing;
 using NSubstitute;
 using Humans.Application.Interfaces;
-using Humans.Application.Services;
+using Humans.Application.Interfaces.Governance;
+using Humans.Application.Services.Governance;
 using Humans.Domain;
 using Humans.Domain.Entities;
 using Humans.Domain.Enums;
