+++ tests/Humans.Application.Tests/Services/CachingApplicationDecisionServiceTests.cs
@@ -4,10 +4,11 @@ using NSubstitute;
 using Humans.Application.DTOs.Governance;
 using Humans.Application.Interfaces;
 using Humans.Application.Interfaces.Caching;
+using Humans.Application.Interfaces.Governance;
 using Humans.Application.Interfaces.Repositories;
 using Humans.Application.Interfaces.Stores;
 using Humans.Domain.Enums;
-using Humans.Infrastructure.Services;
+using Humans.Infrastructure.Services.Governance;
 using Humans.Infrastructure.Stores;
 using Xunit;
 using MemberApplication = Humans.Domain.Entities.Application;
